Changes related to renaming and removed one API
