+++ Ocelot.sln
@@ -98,9 +98,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "OcelotConsulServiceDiscover
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ocelot.Samples.OcelotConsulServiceDiscovery.APIGateway", "Ocelot.Sample.ApiGateway\Ocelot.Samples.OcelotConsulServiceDiscovery.APIGateway.csproj", "{3BC35E7B-91AC-4D0B-9830-598E6E465C90}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ocelot.Samples.OcelotConsulServiceDiscovery.Instance1", "Ocelot.Downstream.Service.Instance1\Ocelot.Samples.OcelotConsulServiceDiscovery.Instance1.csproj", "{3E81C8FA-1E46-44E8-864D-C8ABD21AA8A3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ocelot.Samples.OcelotConsulServiceDiscovery.Instance2", "Ocelot.Downstream.Service.Instance2\Ocelot.Samples.OcelotConsulServiceDiscovery.Instance2.csproj", "{C3EFEB5C-6869-4054-8477-6670AB50474D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ocelot.Samples.OcelotConsulServiceDiscovery.Downstream", "Ocelot.Downstream.Service.Instance1\Ocelot.Samples.OcelotConsulServiceDiscovery.Downstream.csproj", "{3E81C8FA-1E46-44E8-864D-C8ABD21AA8A3}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ocelot.Samples.OcelotConsulServiceDiscovery.Common", "Ocelot.Sample.LoadBalancer.Common\Ocelot.Samples.OcelotConsulServiceDiscovery.Common.csproj", "{6CE1027B-D742-4B5F-AC51-40BE82A7F182}"
 EndProject
@@ -226,10 +224,6 @@ Global
 		{3E81C8FA-1E46-44E8-864D-C8ABD21AA8A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3E81C8FA-1E46-44E8-864D-C8ABD21AA8A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3E81C8FA-1E46-44E8-864D-C8ABD21AA8A3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C3EFEB5C-6869-4054-8477-6670AB50474D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C3EFEB5C-6869-4054-8477-6670AB50474D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C3EFEB5C-6869-4054-8477-6670AB50474D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C3EFEB5C-6869-4054-8477-6670AB50474D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6CE1027B-D742-4B5F-AC51-40BE82A7F182}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6CE1027B-D742-4B5F-AC51-40BE82A7F182}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6CE1027B-D742-4B5F-AC51-40BE82A7F182}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -277,7 +271,6 @@ Global
 		{55B89AB4-D9CA-4B36-8419-F265B62A7D71} = {8FA0CBA0-0338-48EB-B37F-83CA5022237C}
 		{3BC35E7B-91AC-4D0B-9830-598E6E465C90} = {55B89AB4-D9CA-4B36-8419-F265B62A7D71}
 		{3E81C8FA-1E46-44E8-864D-C8ABD21AA8A3} = {55B89AB4-D9CA-4B36-8419-F265B62A7D71}
-		{C3EFEB5C-6869-4054-8477-6670AB50474D} = {55B89AB4-D9CA-4B36-8419-F265B62A7D71}
 		{6CE1027B-D742-4B5F-AC51-40BE82A7F182} = {55B89AB4-D9CA-4B36-8419-F265B62A7D71}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
